# packages/contracts/generated/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "oMLX management API",
    "version": "1"
  },
  "paths": {
...
    "/management/v1/setup": {
      "get": {
        "summary": "Status",
        "operationId": "status_management_v1_setup_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SetupStatus"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create",
        "operationId": "create_management_v1_setup_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetupWrite"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SetupResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SetupResult": {
        "properties": {
          "configured": {
            "type": "boolean",
            "title": "Configured"
          }
        },
        "type": "object",
        "required": [
          "configured"
        ],
        "title": "SetupResult"
      },
      "SetupStatus": {
        "properties": {
          "setup_required": {
            "type": "boolean",
            "title": "Setup Required"
          },
          "allowed": {
            "type": "boolean",
            "title": "Allowed"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          }
        },
        "type": "object",
        "required": [
          "setup_required",
          "allowed",
          "reason"
        ],
        "title": "SetupStatus"
      },
      "SetupWrite": {
        "properties": {
          "key": {
            "type": "string",
            "title": "Key"
          },
          "confirmation": {
            "type": "string",
            "title": "Confirmation"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "key",
          "confirmation"
        ],
        "title": "SetupWrite"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}